server/gateway/requirements.txt
alabaster==0.7.11
alembic==1.0.0
aniso8601==3.0.2
astroid==2.0.4
atomicwrites==1.2.1
attrs==18.2.0
Automat==0.7.0
Babel==2.6.0
certifi==2018.8.24
chardet==3.0.4
click==6.7
confluent-kafka==0.11.5
constantly==15.1.0
docutils==0.14
eyeD3==0.8.7
Flask==1.0.2
Flask-Cors==3.0.7
Flask-JWT-Simple==0.0.3
flask-marshmallow==0.9.0
flask-restplus==0.11.0
Flask-SQLAlchemy==2.3.2
hyperlink==18.0.0
idna==2.7
imagesize==1.1.0
incremental==17.5.0
isort==4.3.4
itsdangerous==0.24
Jinja2==2.10
jsonschema==2.6.0
kazoo==2.5.0
lazy-object-proxy==1.3.1
Mako==1.0.7
MarkupSafe==1.0
marshmallow==2.16.3
marshmallow-sqlalchemy==0.14.1
mccabe==0.6.1
more-itertools==4.3.0
packaging==17.1
pathlib==1.0.1
pluggy==0.8.0
protobuf==3.6.1
psycopg2-binary==2.7.5
py==1.7.0
Pygments==2.2.0
PyHamcrest==1.9.0
PyJWT==1.6.4
pylint==2.1.1
pyparsing==2.2.1
pytest==3.10.0
python-dateutil==2.7.3
python-editor==1.0.3
python-engineio==2.3.2
python-magic==0.4.15
python-socketio==2.0.0
pytz==2018.5
PyYAML==3.13
requests==2.19.1
requests-threads==0.1.1
six==1.11.0
snowballstemmer==1.2.1
Sphinx==1.8.0
sphinxcontrib-websupport==1.1.0
SQLAlchemy==1.2.11
Twisted==18.9.0
typed-ast==1.1.0
urllib3==1.23
Werkzeug==0.14.1
wrapt==1.10.11
zope.interface==4.6.0
